added jwt

+++ server/models/Note.js
@@ -17,20 +17,20 @@ async function createNote(userId, title, content){
 }
 
 //put note/:id
-async function updateNote(id, title, content){
+async function updateNote(userId, id, title, content){
     const result = await db.query(
-        'UPDATE notes SET title = $1, content = $2 WHERE id = $3 RETURNING id, title, content',
-        [title,content,id]
+        'UPDATE notes SET title = $1, content = $2 WHERE id = $3 AND user_id = $4 RETURNING id, title, content',
+        [title,content,id,userId]
     );
     return result.rows[0]
 }
 
 //delete note/:id
 
-async function deleteNote(id){
+async function deleteNote(userId,id){
     const result = await db.query(
-        'DELETE FROM notes WHERE id = $1 RETURNING id, title, content',
-        [id]
+        'DELETE FROM notes WHERE id = $1 AND user_id = $2 RETURNING id, title, content',
+        [id,userId]
     );
 
     return result.rows[0]
